pyproject.toml
[project]
name = "powerplant"
version = "0.46.0"
description = "Deep learning framework for the classification of herbarium samples"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.1",
    "opencv-python>=4.10.0.84",
    "pillow>=11.1.0",
    "tqdm>=4.67.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
exclude = ["/.jj"]

[project.scripts]
powerplant-segment = "powerplant.scripts:segment_cmd"
